api/scripts/requirements.txt
crc8
pyspellchecker
